requirements.txt
pyTelegramBotAPI==4.26.0
pytz==2025.2
requests==2.32.3
APScheduler==3.11.0
google-generativeai==0.8.5